Revert "backend Auth Changes"
This reverts commit b43e1c9c470975bfdfc15cebb566f20a35ce2d29.

+++ migrations/20260519061022_add_company_name_to_stock_items.js
@@ -1,19 +1,27 @@
-/**
- * @param { import("knex").Knex } knex
- * @returns { Promise<void> }
- */
-exports.up = function(knex) {
-  return knex.schema.withSchema('app').table('stock_items', (table) => {
-    table.text('company_name').nullable();
-  });
+export async function up(knex) {
+  const exists = await knex.schema
+    .withSchema("app")
+    .hasColumn("stock_items", "company_name");
+
+  if (!exists) {
+    await knex.schema
+      .withSchema("app")
+      .alterTable("stock_items", function (table) {
+        table.text("company_name").nullable();
+      });
+  }
 };
 
-/**
- * @param { import("knex").Knex } knex
- * @returns { Promise<void> }
- */
-exports.down = function(knex) {
-  return knex.schema.withSchema('app').table('stock_items', (table) => {
-    table.dropColumn('company_name');
-  });
+export async function down(knex) {
+  const exists = await knex.schema
+    .withSchema("app")
+    .hasColumn("stock_items", "company_name");
+
+  if (exists) {
+    await knex.schema
+      .withSchema("app")
+      .alterTable("stock_items", function (table) {
+        table.dropColumn("company_name");
+      });
+  }
 };
